+++ src/main/java/com/knu/ddip/location/application/init/GeoJsonInitializer.java
@@ -1,6 +1,5 @@
 package com.knu.ddip.location.application.init;
 
-import com.knu.ddip.location.application.scheduler.OneTimeRunner;
 import com.knu.ddip.location.application.service.LocationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.ApplicationArguments;
@@ -12,7 +11,7 @@ import org.springframework.stereotype.Component;
 public class GeoJsonInitializer implements ApplicationRunner {
 
     public static final String GEOJSON_INIT_LOCK_KEY = "lock:geojson:init";
-    
+
     private final OneTimeRunner oneTimeRunner;
     private final LocationService locationService;
