+++ src/main/java/com/knu/ddip/location/application/init/OneTimeRunner.java
@@ -1,4 +1,4 @@
-package com.knu.ddip.location.application.scheduler;
+package com.knu.ddip.location.application.init;
 
 public interface OneTimeRunner {
     void runOnce(String lockName, Runnable task);
